requirements.txt

# General
numpy
scipy
pillow>=8.0
img2table>=1.4.0

# Translation
Levenshtein
sacremoses
sentencepiece
transformers

# Box Detection
opencv-contrib-python-headless>=3.0

# Database
databases
sqlalchemy

# OCR
pandas
tesserocr

# Image Manipulation
deskew

# API
fastapi
fastapi-offline
httpx
uvicorn[standard]

# Testing
pytest

# Submitter.py
pdf2image
requests

# Mongo
pymongo==4.3.3

